+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Use interface for default resolver if instantiable
+
 ## v0.2.14 (2023-10-29)
 * Added default resolver interface for nameless resolution
 * Refactored package file structure